third_party/WebKit/LayoutTests/web-animations-api/deprecated-hyphen.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
'use strict';

test(() => {
    var WebAnimationHyphenatedProperty = 1243; // From UseCounter.h
    var target = document.getElementById('target');
    var timing = {fill: 'both'};
    target.animate([{'borderRadius': "1em"}], timing);
    assert_false(internals.isUseCounted(document, WebAnimationHyphenatedProperty));
    target.animate([{'border-radius': "1em"}], timing);
    assert_true(internals.isUseCounted(document, WebAnimationHyphenatedProperty));
}, 'Calling element.animate() with hyphenated property is use counted.');
</script>
