third_party/WebKit/LayoutTests/web-animations-api/deprecated-hyphen-expected.txt

CONSOLE WARNING: line 14: Hyphenated property names in Web Animations keyframes are invalid and therefore ignored. Please use camelCase instead.
This is a testharness.js-based test.
PASS Calling element.animate() with hyphenated property is use counted. 
Harness: the test ran to completion.
